// web/src/http/schemas/webhooks.ts
import { z } from 'zod';

export const webhookListItemSchema = z.object({
  id: z.uuidv7(),
  method: z.string(),
  pathname: z.string(),
  createdAt: z.coerce.date(),
});

export const webhookListSchema = z.object({
  webhooks: z.array(webhookListItemSchema),
  nextCursor: z.string().nullable(),
});

export const webhookDetailsSchema = z.object({
  id: z.uuidv7(),
  method: z.string(),
  pathname: z.string(),
  ip: z.string(),
  statusCode: z.int(),
  contentType: z.string().nullable(),
  contentLength: z.int().nullable(),
  queryParams: z.record(z.string(), z.string()).nullable(),
  headers: z.record(z.string(), z.string()),
  body: z.string(),
  createdAt: z.coerce.date(),
});

export type WebhookListItem = z.infer<typeof webhookListItemSchema>;
export type WebhookList = z.infer<typeof webhookListSchema>;
export type WebhookDetails = z.infer<typeof webhookDetailsSchema>;